+++ assets/javascripts/kitten/components/grid/lego-grid.js
@@ -15,19 +15,15 @@ export const LegoGrid = props => {
   )
 }
 
-LegoGrid.Item = props => {
+LegoGrid.Item = ({ children, ...props }) => {
   const itemClassName = classNames('k-LegoGrid__item', props.className)
 
   return (
-    <div { ...props } className={ itemClassName } />
-  )
-}
-
-LegoGrid.ItemContent = props => {
-  const itemClassName = classNames('k-LegoGrid__item__content', props.className)
-
-  return (
-    <div { ...props } className={ itemClassName } />
+    <div { ...props } className={ itemClassName }>
+      <div className="k-LegoGrid__item__content">
+        { children }
+      </div>
+    </div>
   )
 }
 
@@ -37,9 +33,5 @@ LegoGrid.defaultProps = {
 }
 
 LegoGrid.Item.defaultProps = {
-  className: null,
-}
-
-LegoGrid.ItemContent.defaultProps = {
   className: null,
 }