+++ assets/javascripts/kitten/karl/grid/lego-grid.js
@@ -29,10 +29,8 @@ export const KarlLegoGrid = () => {
           <LegoGrid masonryProps={ masonryProps }>
             { cards.map(card =>
               <LegoGrid.Item key={ card.key }>
-                <LegoGrid.ItemContent>
-                  <div className="karl-ProjectCard"
-                       style={ { height: card.height } } />
-                </LegoGrid.ItemContent>
+                <div className="karl-ProjectCard"
+                     style={ { height: card.height } } />
               </LegoGrid.Item>
             ) }
           </LegoGrid>